% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{chamber_of_secrets}
\alias{chamber_of_secrets}
\title{"Harry Potter and the Chamber of Secrets"}
\format{A character vector with 19 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Chamber_of_Secrets.html}
}
\usage{
chamber_of_secrets
}
\description{
A data set containing the complete text of J.K. Rawling's 1998 novel "Harry
Potter and the Chamber of Secrets". This is the second book in the Harry
Potter series.
}
\keyword{datasets}

