% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{deathly_hallows}
\alias{deathly_hallows}
\title{"Harry Potter and the Deathly Hallows"}
\format{A character vector with 37 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Deathly_Hallows.html}
}
\usage{
deathly_hallows
}
\description{
A data set containing the complete text of J.K. Rawling's 2007 novel "Harry
Potter and the Deathly Hallows". This is the seventh book in the Harry
Potter series.
}
\keyword{datasets}

