% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{goblet_of_fire}
\alias{goblet_of_fire}
\title{"Harry Potter and the Goblet of Fire"}
\format{A character vector with 37 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Goblet_of_Fire.html}
}
\usage{
goblet_of_fire
}
\description{
A data set containing the complete text of J.K. Rawling's 2000 novel "Harry
Potter and the Goblet of Fire". This is the fourth book in the Harry
Potter series.
}
\keyword{datasets}

