% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{half_blood_prince}
\alias{half_blood_prince}
\title{"Harry Potter and the Half-Blood Prince"}
\format{A character vector with 30 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Half-Blood_Prince.html}
}
\usage{
half_blood_prince
}
\description{
A data set containing the complete text of J.K. Rawling's 2005 novel "Harry
Potter and the Half-Blood Prince". This is the sixth book in the Harry
Potter series.
}
\keyword{datasets}

