% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{package}
\name{harrypotter}
\alias{harrypotter}
\alias{harrypotter-package}
\title{J.K. Rowling's Harry Potter Series (Books 1-7)}
\description{
This package contains the complete text of the first seven Harry Potter books,
formatted to be convenient for text analysis.
}

