% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_palette.R
\docType{data}
\name{hp.map}
\alias{hp.map}
\title{Original 'Harry Potter' colour map}
\format{A data frame with 8 variables, one for each movie of the franchise.}
\usage{
hp.map
}
\description{
A dataset containing the averaged RGB values of each frame from every movie of the Harry Potter saga.
}
\details{
Sources: \url{http://movie-colors.com/hp/}..
}
\keyword{datasets}
