% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_palettes}
\alias{hp_palettes}
\title{Available Palettes.}
\format{A list containing all palettes color codes.}
\usage{
hp_palettes
}
\description{
This list contains all the available palettes in the 'harrypotter' package.
}
\keyword{datasets}
