% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{order_of_the_phoenix}
\alias{order_of_the_phoenix}
\title{"Harry Potter and the Order of the Phoenix"}
\format{A character vector with 38 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Order_of_the_Phoenix.html}
}
\usage{
order_of_the_phoenix
}
\description{
A data set containing the complete text of J.K. Rawling's 2003 novel "Harry
Potter and the Order of the Phoenix". This is the fifth book in the Harry
Potter series.
}
\keyword{datasets}

