% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{philosophers_stone}
\alias{philosophers_stone}
\title{"Harry Potter and the Philosopher's Stone"}
\format{A character vector with 17 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Philosophers_Stone.html}
}
\usage{
philosophers_stone
}
\description{
A data set containing the complete text of J.K. Rawling's 1997 novel "Harry
Potter and the Philosopher's Stone". This is the first book in the Harry
Potter series.
}
\keyword{datasets}

