% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{prisoner_of_azkaban}
\alias{prisoner_of_azkaban}
\title{"Harry Potter and the Prisoner of Azkaban"}
\format{A character vector with 22 elements (1 element for each chapter)}
\source{
\url{http://www.readbooksvampire.com/J.K._Rowling/Harry_Potter_and_the_Prisoner_of_Azkaban.html}
}
\usage{
prisoner_of_azkaban
}
\description{
A data set containing the complete text of J.K. Rawling's 1999 novel "Harry
Potter and the Prisoner of Azkaban". This is the third book in the Harry
Potter series.
}
\keyword{datasets}

