% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/hp_palette.R
\docType{data}
\name{hp.map}
\alias{hp.map}
\alias{hp.map}
\title{Original 'Harry Potter' colour map}
\format{A data frame with 4 variables:
\itemize{
  \item V1: Red Value
  \item V2: Green Value
  \item V3: Blue Value
  \item movie: The movie from where the colourmap is extracted
}}
\usage{
hp.map

hp.map
}
\description{
A dataset containing the averaged RGB values of each frame from every movie of the Harry Potter saga.

A dataset containing the averaged RGB values of each frame from every movie of the Harry Potter saga.
}
\details{
Sources: \url{http://movie-colors.com/hp/}..

Sources: \url{http://movie-colors.com/hp/}..
}
\keyword{datasets}
