% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_palette.R
\docType{data}
\name{hp.map}
\alias{hp.map}
\title{Original 'Harry Potter' colour map}
\format{A data frame containing all the colours used in the palette:
\itemize{
#  \item V1: Red value
  \item V2: Green value
  \item V3: Blue value
  \item movie: What refers to the 'movie' of the franchise or the house of Hogwarts. It is intended to be a general option for choosing the specific colour palette.
}}
\usage{
hp.map
}
\description{
A dataset containing the averaged RGB values of each frame from every movie of the Harry Potter saga.
}
\keyword{datasets}
