% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_palette.R
\name{hp_palettes}
\alias{hp_palettes}
\title{Available Palettes.}
\usage{
hp_palettes()
}
\value{
\code{hp_palettes} returns a character vector with the names of the palettes available to use.
}
\description{
This function outpus a vector containing the names of all the available palettes in the 'harrypotter' package.
}
\examples{
hp_palettes()

}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}
}
