% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_palette.R
\name{scale_color_hp}
\alias{scale_color_hp}
\alias{scale_colour_hp}
\alias{scale_color_hp}
\alias{scale_fill_hp}
\title{Harry Potter colour scales}
\usage{
scale_color_hp(..., alpha = 1, begin = 0, end = 1, direction = 1,
  discrete = FALSE, movie = NA, house = "hufflepuff")

scale_colour_hp(..., alpha = 1, begin = 0, end = 1, direction = 1,
  discrete = FALSE, movie = NA, house = "hufflepuff")

scale_fill_hp(..., alpha = 1, begin = 0, end = 1, direction = 1,
  discrete = FALSE, movie = NA, house = "hufflepuff")
}
\arguments{
\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{alpha}{pass through parameter to \code{hp}}

\item{begin}{The (corrected) hue in [0,1] at which the hp colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the hp colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are as output by \link[harrypotter]{hp_pal}. If -1, the order of colors is reversed.}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}

\item{movie}{A number indicating the colormap movie to use. Eight
movies are available: 1,2,3,4,5,6,7 and 8. It is also accepted to desigate
the 8th movie as 7.2 and the 7th movie as 7.1.}

\item{house}{A character string indicating the colourmap from a house to use.
Four houses are available: "Gryffindor", "Slytherin", "Ravenclaw" and "Hufflepuff".}
}
\description{
Uses the Harry Potter color scale.
}
\details{
For \code{discrete == FALSE} (the default) all other arguments are as to
\link[ggplot2]{scale_fill_gradientn} or \link[ggplot2]{scale_color_gradientn}.
Otherwise the function will return a \code{discrete_scale} with the plot-computed
number of colors.
}
\author{
Alejandro Jiménez Rico \email{aljrico@gmail.com}
}
